Move asset IDs out of `ethereumUtils`
