+++ src/__swaps__/screens/Swap/hooks/useSearchCurrencyLists.ts
@@ -16,13 +16,13 @@ import { type AddressOrEth, type AssetType, type ExtendedAnimatedAssetWithColors
 import { type AssetToBuySectionId, type FavoritedAsset, type SearchAsset, type TokenToBuyListItem } from '@/__swaps__/types/search';
 import { type RecentSwap } from '@/__swaps__/types/swap';
 import { analytics } from '@/analytics';
+import { getUniqueId } from '@/entities/assetId';
 import { ChainId } from '@/features/network/types/backendNetworks';
 import { time } from '@/framework/core/utils/time';
 import { isNativeAsset } from '@/handlers/assets';
 import { addHexPrefix } from '@/handlers/web3';
 import { useFavorites } from '@/resources/favorites';
 import { useSwapsStore } from '@/state/swaps/swapsStore';
-import { getUniqueId } from '@/utils/ethereumUtils';
 import isLowerCaseMatch from '@/utils/isLowerCaseMatch';
 import { filterList } from '@/utils/search';
