+++ src/__swaps__/utils/swaps.ts
@@ -15,6 +15,7 @@ import {
 } from '@/__swaps__/screens/Swap/constants';
 import { globalColors } from '@/design-system';
 import { palettes, type ForegroundColor } from '@/design-system/color/palettes';
+import { getUniqueId } from '@/entities/assetId';
 import { IS_TEST } from '@/env';
 import { DEFAULT_SLIPPAGE_BIPS_CHAINID, type RainbowConfig } from '@/features/config/stores/remoteConfig';
 import { ChainId } from '@/features/network/types/backendNetworks';
@@ -35,7 +36,6 @@ import * as i18n from '@/languages';
 import store from '@/redux/store';
 import { useUserAssetsStore } from '@/state/assets/userAssets';
 import { colors } from '@/styles';
-import { getUniqueId } from '@/utils/ethereumUtils';
 import { ETH_ADDRESS, type CrosschainQuote, type Quote, type QuoteError, type QuoteParams } from '@rainbow-me/swaps';
 
 import { swapsStore } from '../../state/swaps/swapsStore';
