+++ src/hooks/useSearchCurrencyList.ts
@@ -5,13 +5,13 @@ import { rankings } from 'match-sorter';
 
 import { useDiscoverSearchQueryStore, useDiscoverSearchStore } from '@/__swaps__/screens/Swap/resources/search/searchV2';
 import { type SearchAsset, type TokenSearchAssetKey, type TokenSearchThreshold } from '@/__swaps__/types/search';
+import { getUniqueId } from '@/entities/assetId';
 import { IS_TEST } from '@/env';
 import { ChainId } from '@/features/network/types/backendNetworks';
 import { addHexPrefix } from '@/handlers/web3';
 import tokenSectionTypes from '@/helpers/tokenSectionTypes';
 import * as i18n from '@/languages';
 import { useFavorites } from '@/resources/favorites';
-import { getUniqueId } from '@/utils/ethereumUtils';
 import { filterList } from '@/utils/search';
 
 import { useTheme } from '../theme/ThemeContext';
