+++ src/systems/funding/execution/depositRapExecution.ts
@@ -2,6 +2,7 @@ import { type Signer } from '@ethersproject/abstract-signer';
 import { ethers, Wallet } from 'ethers';
 
 import { type ParsedAsset } from '@/__swaps__/types/assets';
+import { getUniqueId } from '@/entities/assetId';
 import { resolveManagedExecutionFailure } from '@/features/delegation/utils/managedExecutionFailure';
 import { isInsufficientSponsorBalanceError } from '@/features/delegation/utils/sponsoredCalls';
 import { useBackendNetworksStore } from '@/features/network/stores/backendNetworksStore';
@@ -12,7 +13,6 @@ import { walletExecuteRap } from '@/raps/execute';
 import { rapTypes, type RapSwapActionParameters } from '@/raps/references';
 import erc20ABI from '@/references/erc20-abi.json';
 import { executeFn, Screens, TimeToSignOperation } from '@/state/performance/performance';
-import { getUniqueId } from '@/utils/ethereumUtils';
 import { execute, type PreparedCallsExecution } from '@rainbow-me/sdk';
 import { type CrosschainQuote, type Quote } from '@rainbow-me/swaps';
